backend/requirements.txt
# Hanco-AI Backend Dependencies
# Python 3.11+

# ==================== FastAPI Framework ====================
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# ==================== Firebase ====================
firebase-admin>=6.4.0
google-cloud-firestore>=2.14.0

# ==================== AI Services ====================
google-generativeai>=0.3.0  # Gemini API
openai>=1.10.0  # OpenAI fallback

# ==================== ML/AI ====================
onnxruntime>=1.17.0  # ONNX inference
onnx>=1.15.0  # ONNX model creation
numpy>=1.26.0

# ==================== Web Scraping ====================
crawl4ai>=0.1.0  # Competitor scraping
beautifulsoup4>=4.12.0
lxml>=4.9.0

# ==================== HTTP Clients ====================
httpx>=0.26.0
aiohttp>=3.9.0
requests>=2.31.0

# ==================== Data Processing ====================
pandas>=2.1.0

# ==================== Utilities ====================
python-dotenv>=1.0.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0

# ==================== ML Training ====================
scikit-learn>=1.4.0  # For training pricing model
skl2onnx>=1.16.0  # sklearn to ONNX conversion
# xgboost>=2.0.0  # Optional: alternative algorithm
# lightgbm>=4.2.0  # Optional: alternative algorithm

# ==================== Testing ====================
pytest>=7.4.0
pytest-asyncio>=0.23.0

# ==================== Development ====================
black>=24.0.0
ruff>=0.1.0
mypy>=1.8.0
